+++ CHANGELOG.md
@@ -11,10 +11,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -->
 
 <!-- ## [x.y.z] - YYYY-MM-DD -->
+<!-- ## Unreleased -->
 <!-- ### Changed -->
 <!-- ### Added -->
 <!-- ### Removed -->
 <!-- ### Fixed -->
+
+### Added
+* The `requestUpdateInternal(name, oldValue, options)` method has been added. This method is sometimes useful to call in a custom property setter to optimize performance. It is slightly more efficient than `requestUpdate` since it does not return the `updateComplete` property which can be overridden to do work.
+
+* The protected `performUpdate()` method may now be called to syncronously "flush" a pending update, for example via a property setter. Note, performing a synchronous update only updates the element and not any potentially pending descendants in the element's local DOM ([#959](https://github.com/Polymer/lit-element/issues/959)).
+
+* Constructible stylesheets may now be provided directly as styles, in addition to using the `css` tagged template function ([#853](https://github.com/Polymer/lit-element/issues/853)).
+
+## [2.3.1] - 2020-03-19
+
+### Fixed
+* Add TypeScript type declarations for older versions of TypeScript. We're currently testing back to TS 3.4. We can't commit to never breaking TypeScript builds, but we'll be supporting older versions as best we can.
 
 ## [2.3.0] - 2020-03-18
 
@@ -27,6 +40,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * Added `enableUpdating()` to `UpdatingElement` to enable customizing when updating is enabled [#860](https://github.com/Polymer/lit-element/pull/860).
 * Added `@queryAssignedNodes(slotName, flatten)` decorator to enable querying assignedNodes for a given slot [#860](https://github.com/Polymer/lit-element/pull/860).
 * Added `getStyles()` to `LitElement` to allow hooks into style gathering for component sets [#866](https://github.com/Polymer/lit-element/pull/866).
+* Added `@internalProperty(options)` decorator to define properties internal to an element. [#881](https://github.com/Polymer/lit-element/pull/881).
 
 ### Fixed
 * Ensure `UpdatingElement` allows updates when properties are set after calling `super.update()`.